refactor: rename file
most tests from this file which tested the CLI was moved to test_integration.py
